Replaced the Cognifide logo with Slice icon in package definition
Slice is getting its own logo. I removed the Cognifide logo from the CRX
package definition and replaced it with the new icon that looks like a
slice of pizza (sic!)

Removed the thumbnail/file node, it turns out doing this allows us to get
rid of the OOTB glossy look and feel of icons.